Added block and overlap diagonal
All the elastic models are there and seem to work compatible with the
old implementation. I had also to move the sub_ESH_dns method
to sparsekit because I need it in some el-ph and this would create
circular dependencies. The f90elph_deph test runs with all models and
give identical results (as it should, it's orthogonal and s-like)